requirements.txt
decorator==5.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.1
pillow==11.2.1
proglog==0.1.12
python-dotenv==1.1.1
tqdm==4.67.1
